+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 * Fetch total number of commits from REST API
+* Split repositoriesContributedTo as separate query and treat as optional (i.e., just drop row if it fails)
 
 ### Dependencies